pyproject.toml
[tool.poetry]
name = "film-seeker"
version = "0.1.0"
description = ""
authors = ["d.zaslavskiy <d.zaslavskiy@a1-systems.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
peewee = "^3.17.8"
mysql-connector-python = "^9.1.0"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"
prompt-toolkit = "^3.0.48"
pygments = "^2.19.1"
tabulate = "^0.9.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
